add help pages

+++ commands/clear.js
@@ -10,11 +10,14 @@ module.exports = {
     permission: permissions.moderator,
     code: async (msg, argstring, config) => {
         
-        if (!isint(argstring)) throw errors.syntax;
+        if (!isint(argstring) || argstring > 100) throw errors.syntax;
         
         msg.channel.bulkDelete(argstring).catch(() => {
             throw errors.botperms;
         });
     },
-    help: ``
+    help: `
+    Usage: \`clear [amount]\`
+    
+    Deletes the given amount of messages in the current channel. Thanks to Discord api limitations, you can only clear up to 100 messages at once.`
 }
